requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
marshmallow==3.20.1
pytest==7.4.3
tabulate==0.9.0 

# YC Scraper requirements
pandas>=1.5.0
python-crontab>=2.6.0  # For cron job setup (optional)
pathlib>=1.0.1
selenium==4.15.2  # For Selenium-based web scraping
webdriver-manager==4.0.1  # For automatic chromedriver management